refactor: e2e initialization package
