+++ tests/e2e/docker/image_config.go
@@ -30,8 +30,8 @@ const (
 	previousVersionOsmoRepository = "osmolabs/osmosis-dev"
 	previousVersionOsmoTag        = "v8.0.0-2-debug"
 	// Pre-upgrade repo/tag for osmosis initialization (this should be one version below upgradeVersion)
-	previousVersionInitRepository = "osmolabs/osmosis-init"
-	previousVersionInitTag        = "v8.0.0-4-osmo"
+	previousVersionInitRepository = "osmolabs/osmosis-e2e-init-chain"
+	previousVersionInitTag        = "v8.0.0-rc0"
 	// Hermes repo/version for relayer
 	relayerRepository = "osmolabs/hermes"
 	relayerTag        = "0.13.0"
