+++ tests/e2e/e2e_setup_test.go
@@ -20,8 +20,8 @@ import (
 
 	rpchttp "github.com/tendermint/tendermint/rpc/client/http"
 
-	"github.com/osmosis-labs/osmosis/v9/tests/e2e/chain"
 	dockerconfig "github.com/osmosis-labs/osmosis/v9/tests/e2e/docker"
+	"github.com/osmosis-labs/osmosis/v9/tests/e2e/initialization"
 	"github.com/osmosis-labs/osmosis/v9/tests/e2e/util"
 )
 
@@ -34,7 +34,7 @@ type syncInfo struct {
 }
 
 type validatorConfig struct {
-	validator       chain.Validator
+	validator       initialization.Validator
 	operatorAddress string
 }
 
@@ -49,7 +49,7 @@ type chainConfig struct {
 	skipRunValidatorIndexes map[int]struct{}
 	latestProposalNumber    int
 	latestLockNumber        int
-	meta                    chain.ChainMeta
+	meta                    initialization.ChainMeta
 	validators              []*validatorConfig
 }
 
@@ -70,7 +70,7 @@ const (
 
 var (
 	// whatever number of validator configs get posted here are how many validators that will spawn on chain A and B respectively
-	validatorConfigsChainA = []*chain.ValidatorConfig{
+	validatorConfigsChainA = []*initialization.ValidatorConfig{
 		{
 			Pruning:            "default",
 			PruningKeepRecent:  "0",
@@ -100,7 +100,7 @@ var (
 			SnapshotKeepRecent: 0,
 		},
 	}
-	validatorConfigsChainB = []*chain.ValidatorConfig{
+	validatorConfigsChainB = []*initialization.ValidatorConfig{
 		{
 			Pruning:            "default",
 			PruningKeepRecent:  "0",
@@ -167,11 +167,11 @@ func (s *IntegrationTestSuite) SetupSuite() {
 
 	s.dockerImages = *dockerconfig.NewImageConfig(!skipUpgrade)
 
-	s.configureDockerResources(chain.ChainAID, chain.ChainBID)
-	s.configureChain(chain.ChainAID, validatorConfigsChainA, map[int]struct{}{
+	s.configureDockerResources(initialization.ChainAID, initialization.ChainBID)
+	s.configureChain(initialization.ChainAID, validatorConfigsChainA, map[int]struct{}{
 		3: {}, // skip validator at index 3
 	})
-	s.configureChain(chain.ChainBID, validatorConfigsChainB, map[int]struct{}{})
+	s.configureChain(initialization.ChainBID, validatorConfigsChainB, map[int]struct{}{})
 
 	for i, chainConfig := range s.chainConfigs {
 		s.runValidators(chainConfig, s.dockerImages.OsmosisRepository, s.dockerImages.OsmosisTag, i*10)
@@ -395,7 +395,7 @@ func (s *IntegrationTestSuite) runIBCRelayer(chainA *chainConfig, chainB *chainC
 	s.connectIBCChains(chainA, chainB)
 }
 
-func (s *IntegrationTestSuite) configureChain(chainId string, validatorConfigs []*chain.ValidatorConfig, skipValidatorIndexes map[int]struct{}) {
+func (s *IntegrationTestSuite) configureChain(chainId string, validatorConfigs []*initialization.ValidatorConfig, skipValidatorIndexes map[int]struct{}) {
 	s.T().Logf("starting e2e infrastructure for chain-id: %s", chainId)
 	tmpDir, err := ioutil.TempDir("", "osmosis-e2e-testnet-")
 
@@ -437,7 +437,7 @@ func (s *IntegrationTestSuite) configureChain(chainId string, validatorConfigs [
 
 	fileName := fmt.Sprintf("%v/%v-encode", tmpDir, chainId)
 	s.T().Logf("serialized init file for chain-id %v: %v", chainId, fileName)
-	var initializedChain chain.Chain
+	var initializedChain initialization.Chain
 	// loop through the reading and unmarshaling of the init file a total of maxRetries or until error is nil
 	// without this, test attempts to unmarshal file before docker container is finished writing
 	for i := 0; i < maxRetries; i++ {
@@ -608,10 +608,10 @@ func (s *IntegrationTestSuite) createPreUpgradeState() {
 	chainA := s.chainConfigs[0]
 	chainB := s.chainConfigs[1]
 
-	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, chain.OsmoToken)
-	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, chain.OsmoToken)
-	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, chain.StakeToken)
-	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, chain.StakeToken)
+	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, initialization.OsmoToken)
+	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, initialization.OsmoToken)
+	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, initialization.StakeToken)
+	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, initialization.StakeToken)
 	s.createPool(chainA, "pool1A.json")
 	s.createPool(chainB, "pool1B.json")
 }
@@ -620,10 +620,10 @@ func (s *IntegrationTestSuite) runPostUpgradeTests() {
 	chainA := s.chainConfigs[0]
 	chainB := s.chainConfigs[1]
 
-	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, chain.OsmoToken)
-	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, chain.OsmoToken)
-	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, chain.StakeToken)
-	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, chain.StakeToken)
+	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, initialization.OsmoToken)
+	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, initialization.OsmoToken)
+	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, initialization.StakeToken)
+	s.sendIBC(chainB, chainA, chainA.validators[0].validator.PublicAddress, initialization.StakeToken)
 	s.createPool(chainA, "pool2A.json")
 	s.createPool(chainB, "pool2B.json")
 }
