+++ tests/e2e/e2e_test.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/osmosis-labs/osmosis/v9/tests/e2e/chain"
+	"github.com/osmosis-labs/osmosis/v9/tests/e2e/initialization"
 )
 
 func (s *IntegrationTestSuite) TestIBCTokenTransfer() {
@@ -13,7 +13,7 @@ func (s *IntegrationTestSuite) TestIBCTokenTransfer() {
 	chainB := s.chainConfigs[1]
 	// compare coins of receiver pre and post IBC send
 	// diff should only be the amount sent
-	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, chain.OsmoToken)
+	s.sendIBC(chainA, chainB, chainB.validators[0].validator.PublicAddress, initialization.OsmoToken)
 }
 
 func (s *IntegrationTestSuite) TestSuperfluidVoting() {
