+++ tests/e2e/e2e_util_test.go
@@ -20,7 +20,7 @@ import (
 
 	superfluidtypes "github.com/osmosis-labs/osmosis/v9/x/superfluid/types"
 
-	"github.com/osmosis-labs/osmosis/v9/tests/e2e/chain"
+	"github.com/osmosis-labs/osmosis/v9/tests/e2e/initialization"
 	"github.com/osmosis-labs/osmosis/v9/tests/e2e/util"
 )
 
@@ -129,7 +129,7 @@ func (s *IntegrationTestSuite) sendIBC(srcChain *chainConfig, dstChain *chainCon
 			if ibcCoin.Len() == 1 {
 				tokenPre := balancesBPre.AmountOfNoDenomValidation(ibcCoin[0].Denom)
 				tokenPost := balancesBPost.AmountOfNoDenomValidation(ibcCoin[0].Denom)
-				resPre := chain.OsmoToken.Amount
+				resPre := initialization.OsmoToken.Amount
 				resPost := tokenPost.Sub(tokenPre)
 				return resPost.Uint64() == resPre.Uint64()
 			} else {
