+++ tests/e2e/initialization/chain.go
@@ -1,4 +1,4 @@
-package chain
+package initialization
 
 import (
 	"fmt"
