+++ tests/e2e/initialization/chain/main.go
@@ -7,12 +7,12 @@ import (
 	"os"
 	"time"
 
-	"github.com/osmosis-labs/osmosis/v9/tests/e2e/chain"
+	"github.com/osmosis-labs/osmosis/v9/tests/e2e/initialization"
 )
 
 func main() {
 	var (
-		valConfig    []*chain.ValidatorConfig
+		valConfig    []*initialization.ValidatorConfig
 		dataDir      string
 		chainId      string
 		config       string
@@ -41,7 +41,7 @@ func main() {
 		panic(err)
 	}
 
-	createdChain, err := chain.Init(chainId, dataDir, valConfig, votingPeriod, forkHeight)
+	createdChain, err := initialization.Init(chainId, dataDir, valConfig, votingPeriod, forkHeight)
 	if err != nil {
 		panic(err)
 	}
