+++ tests/e2e/initialization/config.go
@@ -1,4 +1,4 @@
-package chain
+package initialization
 
 import (
 	"encoding/json"
