+++ tests/e2e/initialization/main.go
@@ -1,4 +1,4 @@
-package chain
+package initialization
 
 import "time"
