+++ tests/e2e/initialization/node/main.go
@@ -1,0 +1,7 @@
+package main
+
+// Single-node initialization is not implemented yet This stub is created
+// as an intermediary step before final result.
+func main() {
+	panic("not implemented")
+}
