+++ tests/e2e/initialization/validator.go
@@ -1,4 +1,4 @@
-package chain
+package initialization
 
 import (
 	"encoding/json"
